aws_cdk_python/requirements.txt
-e .
pytest
aws_cdk.aws_dynamodb
aws_cdk.aws_apigateway
